requirements.txt
discord.py[voice]
PyNaCl
youtube_dl
tasks
asyncio
Pillow
yfinance==0.1.63
requests
aiohttp